+++ tokio/src/io/async_fd.rs
@@ -1,5 +1,5 @@
-use crate::io::driver::{Handle, ReadyEvent, Registration};
 use crate::io::Interest;
+use crate::runtime::io::{Handle, ReadyEvent, Registration};
 
 use mio::unix::SourceFd;
 use std::io;
