+++ tokio/src/io/bsd/poll_aio.rs
@@ -1,6 +1,6 @@
 //! Use POSIX AIO futures with Tokio.
 
-use crate::io::driver::{Handle, Interest, ReadyEvent, Registration};
+use crate::runtime::io::{Handle, Interest, ReadyEvent, Registration};
 use mio::event::Source;
 use mio::Registry;
 use mio::Token;
