+++ tokio/src/signal/unix/driver.rs
@@ -2,9 +2,10 @@
 
 //! Signal driver
 
-use crate::io::driver::Driver as IoDriver;
-use crate::io::{Interest, PollEvented};
+use crate::io::PollEvented;
+use crate::io::interest::Interest;
 use crate::park::Park;
+use crate::runtime::io::Driver as IoDriver;
 use crate::signal::registry::globals;
 
 use mio::net::UnixStream;
